refactor: rename components for new cycle countdown component setup
