+++ src/index.ts
@@ -5,7 +5,7 @@ export * from './Checkbox'
 export * from './Collapsible'
 export * from './ColorPicker'
 export * from './Countdown'
-export * from './CountdownNew'
+export * from './CycleCountdown'
 export * from './CycleProgress'
 export * from './DateChanger'
 export * from './Dropdown'
